Separating Binance NTP redux code
