+++ components/brave_new_tab_ui/actions/new_tab_actions.ts
@@ -90,74 +90,6 @@ export const setForegroundStackWidget = (widget: NewTab.StackWidget) => action(t
   widget
 })
 
-export const onBinanceUserTLD = (userTLD: NewTab.BinanceTLD) => action(types.ON_BINANCE_USER_TLD, {
-  userTLD
-})
-
-export const setInitialFiat = (initialFiat: string) => action(types.SET_INITIAL_FIAT, {
-  initialFiat
-})
-
-export const setInitialAmount = (initialAmount: string) => action(types.SET_INITIAL_AMOUNT, {
-  initialAmount
-})
-
-export const setInitialAsset = (initialAsset: string) => action(types.SET_INITIAL_ASSET, {
-  initialAsset
-})
-
-export const setUserTLDAutoSet = () => action(types.SET_USER_TLD_AUTO_SET)
-
 export const setOnlyAnonWallet = (onlyAnonWallet: boolean) => action(types.SET_ONLY_ANON_WALLET, {
   onlyAnonWallet
-})
-
-export const setBinanceSupported = (supported: boolean) => action(types.SET_BINANCE_SUPPORTED, {
-  supported
-})
-
-export const onBinanceClientUrl = (clientUrl: string) => action(types.ON_BINANCE_CLIENT_URL, {
-  clientUrl
-})
-
-export const onValidAuthCode = () => action(types.ON_VALID_AUTH_CODE)
-
-export const setHideBalance = (hide: boolean) => action(types.SET_HIDE_BALANCE, {
-  hide
-})
-
-export const connectToBinance = () => action(types.CONNECT_TO_BINANCE)
-
-export const disconnectBinance = () => action(types.DISCONNECT_BINANCE)
-
-export const onAssetsBalanceInfo = (info: Record<string, Record<string, string>>) => action(types.ON_ASSETS_BALANCE_INFO, {
-  info
-})
-
-export const onAssetDepositInfo = (symbol: string, address: string, tag: string) => action(types.ON_ASSET_DEPOSIT_INFO, {
-  symbol,
-  address,
-  tag
-})
-
-export const onDepositQRForAsset = (asset: string, imageSrc: string) => action(types.ON_DEPOSIT_QR_FOR_ASSET, {
-  asset,
-  imageSrc
-})
-
-export const onConvertableAssets = (convertAsset: string, assets: string[]) => action(types.ON_CONVERTABLE_ASSETS, {
-  convertAsset,
-  assets
-})
-
-export const setDisconnectInProgress = (inProgress: boolean) => action(types.SET_DISCONNECT_IN_PROGRESS, {
-  inProgress
-})
-
-export const setAuthInvalid = (authInvalid: boolean) => action(types.SET_AUTH_INVALID, {
-  authInvalid
-})
-
-export const setSelectedView = (view: string) => action(types.SET_SELECTED_VIEW, {
-  view
 })
