+++ components/brave_new_tab_ui/constants/new_tab_types.ts
@@ -9,6 +9,8 @@ export const enum types {
   NEW_TAB_PREFERENCES_UPDATED = '@@newtab/NEW_TAB_PREFERENCES_UPDATED',
   NEW_TAB_DISMISS_BRANDED_WALLPAPER_NOTIFICATION = '@@newtab/NEW_TAB_DISMISS_BRANDED_WALLPAPER_NOTIFICATION',
   NEW_TAB_SET_INITIAL_DATA = '@@newtab/NEW_TAB_SET_INITIAL_DATA',
+  REMOVE_STACK_WIDGET = '@@newtab/REMOVE_STACK_WIDGET',
+  SET_FOREGROUND_STACK_WIDGET = '@@newtab/SET_FOREGROUND_STACK_WIDGET',
   // Rewards Widget
   CREATE_WALLET = '@@newtab/CREATE_WALLET',
   DISMISS_NOTIFICATION = '@@newtab/DISMISS_NOTIFICATION',
@@ -24,28 +26,7 @@ export const enum types {
   SET_INITIAL_REWARDS_DATA = '@@newtab/SET_INITIAL_REWARDS_DATA',
   SET_PRE_INITIAL_REWARDS_DATA = '@@newtab/SET_PRE_INITIAL_REWARDS_DATA',
   ON_WIDGET_POSITION_CHANGED = '@@newtab/ON_WIDGET_POSITION_CHANGED',
-  SET_ONLY_ANON_WALLET = '@@newtab/SET_ONLY_ANON_WALLET',
-  // Binance Widget
-  ON_BINANCE_USER_TLD = '@@newtab/ON_BINANCE_USER_TLD',
-  SET_INITIAL_ASSET = '@@newtab/SET_INITIAL_ASSET',
-  SET_INITIAL_FIAT = '@@newtab/SET_INITIAL_FIAT',
-  SET_INITIAL_AMOUNT = '@@newtab/SET_INITIAL_AMOUNT',
-  SET_USER_TLD_AUTO_SET = '@@newtab/SET_USER_TLD_AUTO_SET',
-  SET_BINANCE_SUPPORTED = '@@newtab/SET_BINANCE_SUPPORTED',
-  ON_BINANCE_CLIENT_URL = '@@newtab/ON_BINANCE_CLIENT_URL',
-  ON_VALID_AUTH_CODE = '@@newtab/ON_VALID_AUTH_CODE',
-  SET_HIDE_BALANCE = '@@newtab/SET_HIDE_BALANCE',
-  CONNECT_TO_BINANCE = '@@newtab/CONNECT_TO_BINANCE',
-  DISCONNECT_BINANCE = '@@newtab/DISCONNECT_BINANCE',
-  ON_ASSETS_BALANCE_INFO = '@@newtab/ON_ASSETS_BALANCE_INFO',
-  ON_ASSET_DEPOSIT_INFO = '@@newtab/ON_ASSET_DEPOSIT_INFO',
-  ON_DEPOSIT_QR_FOR_ASSET = '@@newtab/ON_DEPOSIT_QR_FOR_ASSET',
-  ON_CONVERTABLE_ASSETS = '@@newtab/ON_CONVERTABLE_ASSETS',
-  SET_DISCONNECT_IN_PROGRESS = '@@newtab/SET_DISCONNECT_IN_PROGRESS',
-  SET_AUTH_INVALID = '@@newtab/SET_AUTH_INVALID',
-  SET_SELECTED_VIEW = '@@newtab/SET_SELECTED_VIEW',
-  REMOVE_STACK_WIDGET = '@@newtab/REMOVE_STACK_WIDGET',
-  SET_FOREGROUND_STACK_WIDGET = '@@newtab/SET_FOREGROUND_STACK_WIDGET'
+  SET_ONLY_ANON_WALLET = '@@newtab/SET_ONLY_ANON_WALLET'
 }
 
 export type DismissBrandedWallpaperNotificationPayload = {
