+++ components/brave_new_tab_ui/containers/newTab/index.tsx
@@ -28,13 +28,14 @@ import { generateQRData } from '../../binance-utils'
 import { SortEnd } from 'react-sortable-hoc'
 import * as newTabActions from '../../actions/new_tab_actions'
 import * as gridSitesActions from '../../actions/grid_sites_actions'
+import * as binanceActions from '../../actions/binance_actions'
 import { getLocale } from '../../../common/locale'
 import currencyData from '../../components/default/binance/data'
 
 interface Props {
   newTabData: NewTab.State
   gridSitesData: NewTab.GridSitesState
-  actions: typeof newTabActions & typeof gridSitesActions
+  actions: typeof newTabActions & typeof gridSitesActions & typeof binanceActions
   saveShowBackgroundImage: (value: boolean) => void
   saveShowClock: (value: boolean) => void
   saveShowTopSites: (value: boolean) => void
